backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-openai
langchain-core
langchain-text-splitters
langchain-chroma
langchain_community
langgraph==0.5.1
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
ormsgpack==1.10.0
xxhash==3.5.0
openai
python-jose
passlib[bcrypt]
pydantic
python-docx
PyPDF2
requests
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
tiktoken
unstructured
python-magic
supabase
langchain-google-genai
tavily-python
google-generativeai